# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Simplon Sénégal",
        "description": "Cet API permet de gérer les candidatures que Simplon Sénégal reçoit pour leurs différentes formations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher la liste de tous les candidats de la plateforme",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Enregistrer un candidat ",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "nom": "string",
                                "prenom": "string",
                                "adresse": "string",
                                "telephone": "string",
                                "email": "string",
                                "password": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{user}": {
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Modifier les infos d'un candidat",
                "operationId": "3e31b0fdd90aa62f59f7de41a72052ea",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "nom": "string",
                                "prenom": "string",
                                "adresse": "string",
                                "telephone": "string",
                                "email": "string",
                                "password": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Connecter un utilisateur",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "email": "string",
                                "password": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh_token": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Regénérer le token",
                "operationId": "85bee97b9fe1e1c2912851cd11b88d6b",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Se déconnecter",
                "operationId": "8992dd4fbd21ecd6bd79f84228dc98cf",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}